tests/submission/W3C/domhtml/HTMLSelectElement07.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - SELECT</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/association-of-controls-and-forms.html#dom-fae-form" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<form action="./files/getData.pl" id="form1" method="post">
<p>
<select dir="ltr" id="selectId" multiple="multiple" name="select1" size="1" tabindex="7">
<option selected="selected" value="EMP1">EMP10001</option>
<option>EMP10002</option>
<option>EMP10003</option>
<option>EMP10004</option>
<option>EMP10005</option>
</select>
</p>
</form>
<p>
<select name="select2">
<option>EMP20001</option>
<option>EMP20002</option>
<option>EMP20003</option>
<option>EMP20004</option>
<option>EMP20005</option>
</select>
</p>
<p>
<select disabled="disabled" name="select3" tabindex="1">
<option>EMP30001</option>
<option>EMP30002</option>
<option>EMP30003</option>
<option>EMP30004</option>
<option>EMP30005</option>
</select>
</p>








<script>
var domtest = async_test("The form attribute returns null if control in not within the context of      a form.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vform;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("select");
      assert_equals(nodeList.length,3);
testNode = nodeList.item(1);
      vform = testNode.form;

      assert_equals(vform, null);
    
    domtest.done();
}));

</script>
<p>The form attribute returns null if control in not within the context of 
    a form. 

    Retrieve the second SELECT element and
    examine its form element.</p>
<div id="log"></div>
</body></html>
